dragon: Move most device-specific files to vendor
It's really nonsense to keep these files in system/
in Pie as we do have a vendor partition. This should
have been done in earlier versions.

This commit moves everything that can be moved without
any extra configuration, and there are some exemptions:

1. rmi4update lies in external/rmi4utils. That is out of
our control and cannot be moved. Luckily it is not essential.

2. Path to enctune.conf is hardcoded in libnvomx.so. Will
`sed` later.

3. Path to bt_vendor.conf can be changed by introducing a
config file. Will do that later.

Change-Id: Ic950484cdaddb94265555c0c3744b2ceaefd5b38